feat: Promote min_stock and purchase quantity to Numeric(10,4)
products.min_stock and purchase_items.quantity were the last two
quantity columns still Integer while every sibling column had already
moved to Numeric(10,4). The business sells half-cartons of eggs, so
integer quantities were silently wrong for both.

Revision 2208c8e60855 adds postgresql_using to all four alter_column
calls (autogenerate omitted them, and Postgres refuses the implicit
integer<->numeric cast without it), and documents in the revision
itself that the downgrade is lossy: Postgres rounds the numeric->integer
cast to the nearest integer (ties away from zero), so any fractional
quantity recorded in the meantime does not survive a downgrade.

+++ app/models/product.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from decimal import Decimal
 from enum import Enum as PyEnum
 from typing import Optional
-from sqlalchemy import Enum as SQLEnum, String, DateTime, Integer, Numeric, Text, Uuid, func, text
+from sqlalchemy import Enum as SQLEnum, String, DateTime, Numeric, Text, Uuid, func, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.database import Base
 
@@ -37,7 +37,9 @@ class Product(Base):
     earning_percent: Mapped[Optional[float]] = mapped_column(Numeric(10, 4), nullable=True)
     earning_fee_amount: Mapped[Optional[float]] = mapped_column(Numeric(12, 2), nullable=True)
     stock: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False, default=0)
-    min_stock: Mapped[int] = mapped_column(Integer, nullable=False, default=0)  # Reorder point
+    min_stock: Mapped[Decimal] = mapped_column(
+        Numeric(10, 4), nullable=False, default=0
+    )  # Reorder point
     status: Mapped[str] = mapped_column(String(20), nullable=False, default="active") # "active" | "inactive"
 
     created_at: Mapped[datetime] = mapped_column(
